requirements.txt
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
cryptography==43.0.0
docutils==0.21.2
idna==3.8
importlib_metadata==8.4.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.0.2
jeepney==0.8.0
keyring==25.3.0
markdown-it-py==3.0.0
mdurl==0.1.2
more-itertools==10.4.0
nh3==0.2.18
numpy==2.1.0
pkginfo==1.10.0
pycparser==2.22
Pygments==2.18.0
python-dotenv==1.0.1
readme_renderer==44.0
requests==2.32.3
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.7.1
SecretStorage==3.3.3
twine==5.1.1
urllib3==2.2.2
zipp==3.20.0
